fix: Move sqliteDb in a native code file
Because the code is based on op-sqlite that is dedicated to
react-native we want to move the related code into a `.native.js` file

By doing this, we ensure that op-sqlite module will be bundled only for
native platforms

On web platforms, the `.js` file will be bundled instead
